third_party/blink/web_tests/compositing/squashing/dont-squash-into-iframes.html

<iframe id="iframe" style="will-change: transform; position: absolute; top: 0px; left: 0px; height: 100px; width: 100px; background: lightblue"></iframe>
<div id="target" style="position: absolute; top: 50px; left: 50px; width: 200px; height: 200px; background: lightgray;"></div>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script>
onload = function() {
    test(function() {
        if (window.internals) {
            var layer_tree = internals.layerTreeAsText(document);
            assert_true(layer_tree.includes("id='iframe'"), 'composited iframe');
            assert_true(layer_tree.includes("id='target'"), 'should not be squashed into iframe');
        }
    });
};
</script>
